+++ docs/components/video/index.tsx
@@ -6,7 +6,7 @@ export function Video({ src }: { src: string }) {
       playsInline
       loop
       controls
-      className="mt-6 rounded-xl border dark:border-zinc-800"
+      className="nextra-focus mt-6 rounded-xl border dark:border-zinc-800"
     >
       <source src={src} type="video/mp4" />
     </video>