+++ packages/nextra-theme-docs/src/components/back-to-top.tsx
@@ -1,13 +1,13 @@
 import cn from 'clsx'
+import { Button } from 'nextra/components'
 import { ArrowRightIcon } from 'nextra/icons'
 import type { ComponentProps, ReactElement } from 'react'
 
 const SCROLL_TO_OPTIONS = { top: 0, behavior: 'smooth' } as const
 
 const scrollToTop: ComponentProps<'button'>['onClick'] = event => {
-  const buttonElement = event.target as HTMLButtonElement
-  const tocElement = buttonElement.parentElement!
-    .parentElement as HTMLDivElement
+  const buttonElement = event.currentTarget
+  const tocElement = buttonElement.parentElement!.parentElement!
 
   window.scrollTo(SCROLL_TO_OPTIONS)
   tocElement.scrollTo(SCROLL_TO_OPTIONS)
@@ -25,17 +25,23 @@ export function BackToTop({
   hidden: boolean
 }): ReactElement {
   return (
-    <button
+    <Button
       aria-hidden="true"
       onClick={scrollToTop}
       disabled={hidden}
-      className={cn(
-        '_flex _items-center _gap-1.5 _transition _opacity-100 disabled:_opacity-0',
-        className
-      )}
+      className={({ disabled }) =>
+        cn(
+          '_flex _items-center _gap-1.5',
+          disabled ? '_opacity-0' : '_opacity-100',
+          className
+        )
+      }
     >
       Scroll to top
-      <ArrowRightIcon className="_h-4 _-rotate-90 _border _rounded-full _border-current" />
-    </button>
+      <ArrowRightIcon
+        height="16"
+        className="_-rotate-90 _border _rounded-full _border-current"
+      />
+    </Button>
   )
 }