+++ packages/nextra-theme-docs/src/components/search.tsx
@@ -15,7 +15,6 @@ import { Fragment, useCallback, useEffect, useRef, useState } from 'react'
 import { useMenu, useThemeConfig } from '../contexts'
 import type { SearchResult } from '../types'
 import { renderComponent, renderString } from '../utils'
-import { Input } from './input'
 
 type SearchProps = {
   className?: string
@@ -133,36 +132,62 @@ export function Search({
   // }, [results])
 
   return (
-    <Combobox
-      as="div"
-      // value={selected}
-      onChange={handleSelect}
-      className={cn('nextra-search _relative md:_w-64', className)}
-    >
-      <ComboboxInput
-        ref={inputRef}
-        as={Input}
-        autoComplete="off"
-        value={value}
-        onChange={handleChange}
-        onFocus={handleFocus}
-        onBlur={handleFocus}
-        type="search"
-        placeholder={renderString(themeConfig.search.placeholder)}
-        suffix={icon}
-      />
+    <Combobox onChange={handleSelect}>
+      <div
+        className={cn(
+          '_not-prose', // for blog
+          '_relative _flex _items-center',
+          '_text-gray-900 dark:_text-gray-300',
+          'contrast-more:_text-gray-800 contrast-more:dark:_text-gray-300',
+          className
+        )}
+      >
+        <ComboboxInput
+          ref={inputRef}
+          spellCheck={false}
+          className={({ focus }) =>
+            cn(
+              '_rounded-lg _px-3 _py-2 _transition-colors',
+              '_w-full md:_w-64',
+              '_text-base _leading-tight md:_text-sm',
+              focus
+                ? '_bg-transparent nextra-focusable'
+                : '_bg-black/[.05] dark:_bg-gray-50/10',
+              'placeholder:_text-gray-500 dark:placeholder:_text-gray-400',
+              'contrast-more:_border contrast-more:_border-current',
+              '[&::-webkit-search-cancel-button]:_appearance-none'
+            )
+          }
+          autoComplete="off"
+          type="search"
+          onChange={handleChange}
+          onFocus={handleFocus}
+          onBlur={handleFocus}
+          value={value}
+          placeholder={renderString(themeConfig.search.placeholder)}
+        />
+        {icon}
+      </div>
       <ComboboxOptions
         transition
         anchor={{ to: 'top end', gap: 10, padding: 16 }}
-        className={cn(
-          'nextra-search-results nextra-scrollbar max-md:_h-full',
-          error || loading || !results.length
-            ? 'md:_h-[100px]'
-            : // headlessui adds max-height as style, use !important to override
-              'md:!_max-h-[min(calc(100vh-5rem),400px)]',
-          '_w-[--input-width] md:_w-[576px]',
-          'empty:_invisible'
-        )}
+        className={({ open }) =>
+          cn(
+            'nextra-scrollbar max-md:_h-full',
+            '_border _border-gray-200 _text-gray-100 dark:_border-neutral-800',
+            '_z-20 _rounded-xl _py-2.5 _shadow-xl',
+            'contrast-more:_border contrast-more:_border-gray-900 contrast-more:dark:_border-gray-50',
+            '_backdrop-blur-lg _bg-[rgb(var(--nextra-bg),.8)]',
+            'motion-reduce:_transition-none _transition-opacity',
+            open ? '_opacity-100' : '_opacity-0',
+            error || loading || !results.length
+              ? 'md:_h-[100px]'
+              : // headlessui adds max-height as style, use !important to override
+                'md:!_max-h-[min(calc(100vh-5rem),400px)]',
+            '_w-full md:_w-[576px]',
+            'empty:_invisible'
+          )
+        }
       >
         {error ? (
           <span className="_flex _select-none _justify-center _gap-2 _p-8 _text-center _text-sm _text-red-500">
@@ -182,13 +207,16 @@ export function Search({
                 as={NextLink}
                 value={searchResult}
                 href={searchResult.route}
-                className={cn(
-                  '_mx-2.5 _break-words _rounded-md',
-                  'contrast-more:_border',
-                  'data-[focus]:_ring data-[focus]:_bg-primary-500/10 data-[focus]:!_text-primary-600 data-[focus]:contrast-more:_border-primary-500',
-                  '_text-gray-800 contrast-more:_border-transparent dark:_text-gray-300',
-                  '_block _scroll-m-12 _px-2.5 _py-2'
-                )}
+                className={({ focus }) =>
+                  cn(
+                    '_mx-2.5 _break-words _rounded-md',
+                    'contrast-more:_border',
+                    focus
+                      ? '_text-primary-600 contrast-more:_border-current _bg-primary-500/10'
+                      : '_text-gray-800 dark:_text-gray-300 contrast-more:_border-transparent',
+                    '_block _scroll-m-12 _px-2.5 _py-2'
+                  )
+                }
               >
                 {searchResult.children}
               </ComboboxOption>